+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* TODO: A handy command-line tool to help you identify and interactively remove unused packages from your Dart/Flutter project's `pubspec.yaml`.
+